% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{messy}
\alias{messy}
\title{Simulated Data}
\format{A data frame with 1000 rows and 3 variables}
\description{
Simulated (messy) data set to help illustrate some of the uses of basic
pattern analysis.
}
\details{
\itemize{
  \item \code{Gender} Gender in various formats.
  \item \code{Date} Dates in various formats.
  \item{Phone} Phone numbers in various formats.
}
}
\examples{
data(messy)
bpa(messy, unique_only = TRUE, ws_char = " ")
}
\keyword{datasets}

