% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_ws.R
\name{trim_ws}
\alias{trim_ws}
\title{Trim Leading and Trailing Whitespace}
\usage{
trim_ws(x, ...)
}
\arguments{
\item{x}{A data frame or vector.}

\item{...}{Additional optional arguments to be passed onto
\code{\link{trimws}}.}
}
\description{
\code{trim_ws} trims leading and trailing whitespace from data.
}
\examples{
# Toy example
d <- data.frame(x = c(" a ", "b ", "c"),
                y = c("   1 ", "2", " 3"),
                z = c(4, 5, 6))
print(d)  # print data as is
trim_ws(d)  # print data with whitespace trimmed off
sapply(trim_ws(d), class)  # check that column types are preserved
}

