% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_pattern_analysis.R
\name{basic_pattern_analysis}
\alias{basic_pattern_analysis}
\alias{basic_pattern_analysis.data.frame}
\alias{basic_pattern_analysis.default}
\alias{bpa}
\title{Basic Pattern Analysis}
\usage{
basic_pattern_analysis(x, unique_only = FALSE, show_ws = TRUE,
  useNA = c("no", "ifany", "always"), ...)

\method{basic_pattern_analysis}{default}(x, unique_only = FALSE,
  show_ws = TRUE, useNA = c("no", "ifany", "always"), ...)

\method{basic_pattern_analysis}{data.frame}(x, unique_only = FALSE,
  show_ws = TRUE, useNA = c("no", "ifany", "always"), ...)

bpa(x, ...)
}
\arguments{
\item{x}{A data frame or character vector.}

\item{unique_only}{Logical indicating whether or not to only show the unique
patterns. Default is \code{TRUE}.}

\item{show_ws}{Logical indicating whether or not to show whitespace
as a \code{"w"} character. Default is \code{TRUE}.}

\item{useNA}{Logical indicating whether to include \code{NA} values in the
table. See \code{\link{table}} for details.}

\item{...}{Additional optional arguments to be passed onto \code{llply}.}
}
\description{
Perform a basic pattern analysis
}
\examples{
basic_pattern_analysis(iris)
basic_pattern_analysis(iris, unique_only = TRUE)
}

